Fix things mentioned during code review
Ref: https://github.com/dolphin-emu/dolphin/pull/4917
